+++ CHANGELOG.md
@@ -6,6 +6,10 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-7491 Fixed column labels overflow in harvesting progress table
+* VFS-7228 Better UX of token consumer page
+* VFS-7516 Added new file details to harvester indices: "file type" and "dataset info"
+* VFS-7401 Updated commons
 * VFS-7477 Fixed removing space from harvester
 * VFS-6779 Added links to harvesters/spaces from list of harvesters/spaces
 * VFS-7413 Fixed missing members with duplicated names